// packages/optimizely-configure-commerce-graph-client/src/client/types.ts
export enum SchemaVersion {
  /**
   * This is the default (current) schema of Optimizely Graph.
   */
  Default = "current",
  /**
   * This is a future schema, which has been staged to to be released in the
   * future.
   */
  Next = "new",
}

export enum OptiCmsSchema {
  CMS12 = "OPTI-CMS-12",
  CMS13 = "OPTI-CMS-13",
}